requirements.txt
altgraph==0.17.5
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
